requirements.txt
Django~=4.0.1
redis~=4.1.0
django-adminlte3~=0.1.6
django-storages~=1.12.3
requests~=2.27.1
django-redis~=5.2.0
django-thumbnails
django-s3-storage
sorl-thumbnail
django-import-export
django-cors-headers
agora-token-builder
gunicorn
uwsgi
pymysql